requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-dotenv==1.0.1
google-generativeai==0.8.6
pydantic==2.12.5
requests==2.33.0
solana==0.36.0
solders==0.21.0
anchorpy==0.21.0
sqlalchemy==2.0.38
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic-settings==2.8.1
aiofiles==24.1.0
python-multipart==0.0.20
